make the exclusion regexes slice atomic
